break singple post down into separate components
